Move most static methods in FileLocator back to SIL.Core
