+++ SIL.Windows.Forms.GeckoBrowserAdapter/ReflectiveGeckoBrowser.cs
@@ -86,11 +86,11 @@ namespace SIL.Windows.Forms.GeckoBrowserAdapter
 			string xulRunnerPath = Environment.GetEnvironmentVariable("XULRUNNER");
 			if (!Directory.Exists(xulRunnerPath))
 			{
-				xulRunnerPath = Path.Combine(FileLocator.DirectoryOfApplicationOrSolution, "xulrunner");
+				xulRunnerPath = Path.Combine(FileLocationUtilities.DirectoryOfApplicationOrSolution, "xulrunner");
 				if (!Directory.Exists(xulRunnerPath))
 				{
 					// if this is a programmer, go look in the lib directory
-					xulRunnerPath = Path.Combine(FileLocator.DirectoryOfApplicationOrSolution,
+					xulRunnerPath = Path.Combine(FileLocationUtilities.DirectoryOfApplicationOrSolution,
 						Path.Combine("lib", "xulrunner"));
 
 					// on my build machine, I really like to have the dir labelled with the version.
@@ -98,7 +98,7 @@ namespace SIL.Windows.Forms.GeckoBrowserAdapter
 					// so we only use it if we don't find the unnumbered alternative.
 					if (!Directory.Exists(xulRunnerPath))
 					{
-						xulRunnerPath = Path.Combine(FileLocator.DirectoryOfApplicationOrSolution,
+						xulRunnerPath = Path.Combine(FileLocationUtilities.DirectoryOfApplicationOrSolution,
 							Path.Combine("lib", "xulrunner" + XulRunnerVersion));
 					}
